requirements.txt
argh==0.26.2
git-gifi==0.12
gitdb==0.6.4
GitPython==1.0.1
pathtools==0.1.2
PyGithub==1.25.2
PyYAML==3.13
six==1.12.0
slackclient==0.15
smmap==0.9.0
watchdog==0.9.0
websocket-client==0.54.0
